Added DevTools and DBConnection
Connected the server to the DB (created a DevDB) and entered the session
configuration information (the String of the secret and two boolean
options) there and tested reading that so we can start loading configs
from the DB if present, from the locals if not. Also added some dev
tools to make it easy to turn on and off messages to the console for
development purposes, all controllable by the 'dev' console parameter at
runtime.